* apache-github/trunk: (148 commits)
  MINOR: remove NewTopic#NO_PARTITIONS and NewTopic#NO_REPLICATION_FACTOR as they are duplicate to CreateTopicsRequest#NO_NUM_PARTITIONS and CreateTopicsRequest#NO_REPLICATION_FACTOR (#9077)
  MINOR: Remove staticmethod tag to be able to use logger of instance (#9086)
  MINOR: Adjust 'release.py' script to use shell when using gradlewAll and PGP signing, which were required to build the 2.6.0 RCs (#9045)
  MINOR: Update dependencies for Kafka 2.7 (part 1) (#9082)
  MINOR: INFO log4j when request re-join (#9068)
  MINOR: Recommend Java 11 (#9080)
  KAFKA-10306: GlobalThread should fail on InvalidOffsetException (#9075)
  KAFKA-10158: Fix flaky testDescribeUnderReplicatedPartitionsWhenReassignmentIsInProgress (#9022)
  MINOR: code cleanup for `VOut` inconsistent naming (#8907)
  KAFKA-10246 : AbstractProcessorContext topic() throws NPE (#9034)
  KAFKA-10305: Print usage when parsing fails for ConsumerPerformance (#9071)
  MINOR: removed incorrect deprecation annotations (#9061)
  MINOR: speed up release script (#9070)
  MINOR: add task ':streams:testAll' (#9073)
  KAFKA-10301: Do not clear Partition#remoteReplicasMap during partition assignment updates (#9065)
  KAFKA-10268: dynamic config like "--delete-config log.retention.ms" doesn't work (#9051)
  KAFKA-10300 fix flaky core/group_mode_transactions_test.py (#9059)
  MINOR: Publish metrics package in the javadoc (#9036)
  KAFKA-8264: decrease the record size for flaky test
  KAFKA-5876: Add new exception types for Interactive Queries (#8200)
  ...

+++ clients/src/main/java/org/apache/kafka/common/security/authenticator/SaslClientAuthenticator.java
@@ -676,7 +676,7 @@ public class SaslClientAuthenticator implements Authenticator {
                 double pctWindowJitterToAvoidReauthenticationStormAcrossManyChannelsSimultaneously = 0.10;
                 double pctToUse = pctWindowFactorToTakeNetworkLatencyAndClockDriftIntoAccount + RNG.nextDouble()
                         * pctWindowJitterToAvoidReauthenticationStormAcrossManyChannelsSimultaneously;
-                sessionLifetimeMsToUse = (long) (positiveSessionLifetimeMs.longValue() * pctToUse);
+                sessionLifetimeMsToUse = (long) (positiveSessionLifetimeMs * pctToUse);
                 clientSessionReauthenticationTimeNanos = authenticationEndNanos + 1000 * 1000 * sessionLifetimeMsToUse;
                 log.debug(
                         "Finished {} with session expiration in {} ms and session re-authentication on or after {} ms",
@@ -688,7 +688,7 @@ public class SaslClientAuthenticator implements Authenticator {
 
         public Long reauthenticationLatencyMs() {
             return reauthenticating()
-                    ? Long.valueOf(Math.round((authenticationEndNanos - reauthenticationBeginNanos) / 1000.0 / 1000.0))
+                    ? Math.round((authenticationEndNanos - reauthenticationBeginNanos) / 1000.0 / 1000.0)
                     : null;
         }
